exercises comptime & builtin

+++ exercises/060_floats.zig
@@ -40,7 +40,7 @@ pub fn main() void {
     //
     // We'll convert this weight from tons to kilograms at a
     // conversion of 907.18kg to the ton.
-    var shuttle_weight: f64 = 907.18 * 2.2e+3;
+    var shuttle_weight: f32 = 907.18 * 2.2e+3;
 
     // By default, float values are formatted in scientific
     // notation. Try experimenting with '{d}' and '{d:.3}' to see
